pyproject.toml
[project]
name = "synconf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "richarddwang", email = "richardyy1188@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "docstring-parser>=0.17.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = ["ipykernel>=6.30.1", "pytest>=8.4.1", "ruff>=0.12.7"]

[project.urls]
Homepage = "https://github.com/richarddwang/synconf"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[[tool.uv.index]]
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "ailab-gitlab"
url = "http://10.11.60.1:8287/api/v4/groups/22/-/packages/pypi/simple"
